+++ plugins/module_utils/client.py
@@ -8,7 +8,7 @@ import codecs
 import json
 import traceback
 
-from ansible.module_utils._text import to_text
+from ansible.module_utils.common.text.converters import to_text
 from urllib.parse import quote
 from ansible.module_utils.urls import fetch_url