+++ plugins/modules/ping.py
@@ -23,7 +23,8 @@ author:
   - Shiva Shankar Vaddepally (@shivashankar-vaddepally)
 options:
   state:
-    choices: []
+    choices:
+      - present
     default: present
     description:
       - The state of the resource being configured by the module on the NetScaler
@@ -39,19 +40,22 @@ options:
         in the resource.
       - If no, the module will return error if any non-updatable parameters are provided.
     type: str
-  I:
+  interface:
     type: str
     description:
       - Network interface on which to ping, if you have multiple interfaces.
-  S:
+      - 'Deprecated alias: C(I).'
+  source_ip:
     type: str
     description:
       - Source IP address to be used in the outgoing query packets. If the IP addrESS
         does not belongs to this appliance, an error is returned and nothing is sent.
-  T:
+      - 'Deprecated alias: C(S).'
+  traffic_domain:
     type: int
     description:
       - Traffic Domain Id
+      - 'Deprecated alias: C(T).'
   c:
     type: int
     description:
@@ -61,10 +65,11 @@ options:
     type: str
     description:
       - Address of host to ping.
-  i:
+  interval:
     type: int
     description:
       - Waiting time, in seconds. The default value is 1 second.
+      - 'Deprecated alias: C(i).'
   n:
     type: bool
     description:
@@ -79,19 +84,41 @@ options:
     description:
       - Quiet output. Only the summary is printed. For Nitro API, this flag is set
         by default.
-  s:
+  packet_size:
     type: int
     description:
       - Data size, in bytes. The default value is 56.
-  t:
+      - 'Deprecated alias: C(s).'
+  timeout:
     type: int
     description:
       - Time-out, in seconds, before ping exits.
+      - 'Deprecated alias: C(t).'
 extends_documentation_fragment: netscaler.adc.netscaler_adc
 
 """
 
 EXAMPLES = r"""
+---
+- name: Sample ping playbook
+  hosts: demo_netscalers
+  gather_facts: false
+  tasks:
+    - name: Ping a host from the NetScaler ADC
+      delegate_to: localhost
+      register: ping_result
+      netscaler.adc.ping:
+        state: present
+        hostName: 127.0.0.1
+        c: 4                # number of packets to send
+        packet_size: 56     # data size in bytes (deprecated alias: s)
+        interval: 1         # seconds between packets (deprecated alias: i)
+        timeout: 5          # seconds before ping exits (deprecated alias: t)
+
+    - name: Show the ping command output
+      delegate_to: localhost
+      ansible.builtin.debug:
+        msg: "{{ ping_result.ping.response }}"
 """
 
 RETURN = r"""
@@ -101,6 +128,13 @@ changed:
   returned: always
   type: bool
   sample: true
+ping:
+  description: Result of the ping command. The command output is in the C(response) field.
+  returned: on success (not in check mode)
+  type: dict
+  sample: {'response': 'PING 127.0.0.1 (127.0.0.1): 56 data bytes\n64 bytes from 127.0.0.1:
+      icmp_seq=0 ttl=64 time=0.045 ms\n\n--- 127.0.0.1 ping statistics ---\n3 packets
+      transmitted, 3 packets received, 0.0% packet loss'}
 diff:
   description: Dictionary of before and after changes
   returned: always